pyproject.toml
[tool.poetry]
name = "jaya"
version = "0.1.0"
description = ""
authors = ["Yago Ferreira <yagojj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
motor = "^3.4.0"
python-decouple = "^3.8"
meeseeks-singleton = "^0.4.4"
witch-doctor = "^1.1.0"
loglifos = "^0.2.0"
aiohttp = "^3.9.5"
pre-commit = "^3.7.0"
fastapi = "^0.110.2"
uvicorn = "^0.29.0"


[tool.poetry.group.dev.dependencies]
pyfiglet = "^1.0.2"
pre-commit = "^3.7.0"
pylint = "^3.1.0"
pytest = "^8.1.1"
black = "^24.4.0"
pytest-asyncio = "^0.23.6"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
